+++ se_app/src/main/java/org/group35/view/AIDialogController.java
@@ -14,7 +14,8 @@ import org.group35.model.User;
 import org.group35.runtime.ApplicationRuntime;
 import java.time.format.DateTimeFormatter;
 import javafx.application.Platform;
-import org.group35.service.LocalizationService;
+import org.group35.util.TimezoneUtils;
+import org.group35.util.CurrencyUtils;
 
 public class AIDialogController {
     @FXML
@@ -150,7 +151,7 @@ public class AIDialogController {
      */
     private void loadLocalInfo() {
         // 异步获取本地化信息，避免阻塞UI
-        LocalizationService.getCurrentLocalInfoAsync()
+        TimezoneUtils.getCurrentLocalInfoAsync()
                 .thenAccept(localInfo -> {
                     // 在JavaFX主线程中更新UI
                     Platform.runLater(() -> updateLocalInfoDisplay(localInfo));
@@ -158,7 +159,7 @@ public class AIDialogController {
                 .exceptionally(throwable -> {
                     System.err.println("Error loading local info: " + throwable.getMessage());
                     // 使用默认信息
-                    Platform.runLater(() -> updateLocalInfoDisplay(LocalizationService.getCurrentLocalInfo()));
+                    Platform.runLater(() -> updateLocalInfoDisplay(TimezoneUtils.getCurrentLocalInfo()));
                     return null;
                 });
     }
@@ -166,14 +167,14 @@ public class AIDialogController {
     /**
      * 新增方法：更新本地化信息显示
      */
-    private void updateLocalInfoDisplay(LocalizationService.LocalInfo localInfo) {
+    private void updateLocalInfoDisplay(TimezoneUtils.LocalInfo localInfo) {
         try {
             if (timezoneLabel != null) {
                 timezoneLabel.setText(localInfo.getTimezone());
             }
 
             if (currencyLabel != null) {
-                String formattedCurrency = LocalizationService.formatCurrencyName(localInfo.getCurrency());
+                String formattedCurrency = CurrencyUtils.formatCurrencyName(localInfo.getCurrency());
                 currencyLabel.setText(formattedCurrency);
             }
 
@@ -199,7 +200,7 @@ public class AIDialogController {
      */
     @FXML
     private void refreshLocalInfo() {
-        LocalizationService.refreshCache();
+        TimezoneUtils.refreshCache();
         loadLocalInfo();
     }
