--- se_app/src/main/resources/org/group35/view/config.properties
@@ -1,15 +1,0 @@
-# GeoNames API Configuration
-geonames.username=Dusk_0027
-geonames.search.url=http://api.geonames.org/searchJSON
-geonames.timezone.url=http://api.geonames.org/timezoneJSON
-
-# Exchange Rate API Configuration
-exchange.api.url=https://api.exchangerate-api.com/v4/latest/USD
-
-# Default fallback settings (?????????)
-default.location=Unknown
-default.currency=USD
-default.timezone=UTC
-
-# Update Interval (minutes)
-update.interval=30
